Added analysis to look at weak proxies X Z in UKBB
